+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.6
+
+- Fix onFaceInsideBox still called when required center but face is not centered
+- Add onFaceOutsideBox callback
+
 ## 0.0.5
 
 - Change Face Box Painter